Rework goals page
